Add files via upload
WF_v1d.py is a python program which can be used by those learning the German language mainly through English and Greek. Type a German word at the top field to find the meaning by pressing one of the buttons below (e.g. Langescheidt). If the definition exists, it will appear in the box below. If your press the long button, it definition will be transferred to the third box, which is designed to host all definitions. You can then save those definitions to create a text file with all the words you searched.
If you need to see how a word is used, i.e. examples, you type the word in the top field, just as before and then click one of the buttons at the bottom (e.g. WordRef, Collins, Glosbe)